requirements.txt
django==5.1.1
djangorestframework==3.15.2
pytest-django==4.9.0
model-bakery==1.19.5
Faker==30.8.0
openpyxl==3.1.5
python-docx==1.1.2
Pillow==11.3.0
gunicorn==22.0.0